Change data structure
Now use the flatsurf convention for half-edges, ie the edge
permutation is implicit and always of the form (0,1)(2,3)...
In the case of folded edges, we use partial permutations. Namely
if i is a folded edge (2i) is a fixed point of the edge permutation
and 2i+1 not in the domain. This can be checked by looking at the
vertex or face permutation.